requirements.txt
flake8
pep8
autopep8
numpy
pillow-SIMD
pyqt5
